python/requirements.txt
requests>=2.28.0,<3.0
python-dotenv>=1.0,<2.0
# pydantic[email]>=1.10,<3.0
